Fixed a number of unittests

+++ lib/gssapi/krb5/test_cfx.c
@@ -160,7 +160,7 @@ main(int argc, char **argv)
     test_special(context, crypto, 1, 60);
     test_special(context, crypto, 0, 60);
 
-    for (i = 0; i < sizeof(tests)/sizeof(tests[0]); i++) {
+    for (i = 0; i < (int) (sizeof(tests)/sizeof(tests[0])); i++) {
 	test_range(&tests[i], 1, context, crypto);
 	test_range(&tests[i], 0, context, crypto);
     }
