+++ lib/krb5/test_crypto_wrapping.c
@@ -156,7 +156,7 @@ main(int argc, char **argv)
     if (ret)
 	errx (1, "krb5_init_context failed: %d", ret);
 
-    for (i = 0; i < sizeof(enctypes)/sizeof(enctypes[0]); i++) {
+    for (i = 0; i < (int) sizeof(enctypes)/sizeof(enctypes[0]); i++) {
 	krb5_enctype_enable(context, enctypes[i]);
 
 	test_wrapping(context, 0, 1024, 1, enctypes[i]);
