+++ lib/krb5/test_pkinit_dh2key.c
@@ -205,7 +205,7 @@ main(int argc, char **argv)
     if (ret)
 	errx (1, "krb5_init_context failed: %d", ret);
 
-    for (i = 0; i < sizeof(tests)/sizeof(tests[0]); i++) {
+    for (i = 0; i < (int) (sizeof(tests)/sizeof(tests[0])); i++) {
 	test_dh2key(i, context, &tests[i].X, NULL, NULL,
 		    tests[i].type, &tests[i].key);
     }
